refactor: action, render 클래스 모듈화
